Move non-English functions to repo with translations
